requirement.txt
matplotlib==3.3.3
pandas==1.1.5
tensorflow-gpu==2.4.1
xlrd==1.2.0
scipy==1.5.4
openpyxl==3.0.5

